Allow loading from system and local folders l10n.
Use gotext to load both from system, and any optional local folder l10n
content as override.
On system, we are using the main known system folders, including the
langpack ones.
If a fs.FS is provided, we will load the current language if we find any
to override the default domain with it.

The fs.FS can be embedded translations when running from the repository
for instance.