ci: Automatically cancel in-progress workflow runs on push
If commits are pushed to the same PR in quick succession, the CI jobs
run multiple times in parallel, even though older runs are usually not
needed. This change saves some resources by automatically cancelling
in-progress workflows on new pushes to the same branch.

The implementation is directly from the example of the concurrency keyword:
https://docs.github.com/en/actions/reference/workflows-and-actions/workflow-syntax#example-using-concurrency-and-the-default-behavior